# openapi.yaml
openapi: 3.1.0
info:
  title: webhook-out
  version: "1.0.0"
  description: |
    Outbound webhooks: subscriptions with event patterns, fan-out on publish, HMAC-signed
    deliveries with scheduled retries, secret rotation, delivery history.

    ## Authentication and roles

    Every `/v1/*` route and `/metrics` require `Authorization: Bearer <secret>`. Each configured
    API key carries exactly one role: `read`, `write`, `publish` or `readwrite`. Routes are gated
    on a *need* (`read`, `write` or `publish`), and a key's role must be in that need's grant list:

    | Need     | Roles that satisfy it              |
    |----------|-------------------------------------|
    | `read`   | `read`, `readwrite`                  |
    | `write`  | `write`, `readwrite`                 |
    | `publish`| `publish`, `write`, `readwrite`      |

    This lattice is asymmetric and is the most distinctive part of this service's authorization
    model: a `write` key also satisfies `publish` (it can call `POST /v1/events`), but a
    `publish`-only key satisfies *only* the `publish` need. It does **not** satisfy `read` or
    `write`, so a `publish`-only key can call `POST /v1/events` and nothing else in this API — every
    other route (including plain reads like `GET /v1/subscriptions`) responds `403 FORBIDDEN` to it.
    A `readwrite` key satisfies every need.

    A missing or unrecognized key gets `401 UNAUTHORIZED`. A recognized key whose role does not
...

...

...
    e.g. `order.paid`.

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/subscriptions:
    post:
      operationId: webhookOut.subscriptions.create
      summary: Create a subscription
      description: |
        Creates a subscription and returns its brand-new signing secret. **The secret is shown
        only here and in the `rotate` response — it cannot be retrieved again afterward** (only a
        sealed form is stored).
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionBody'
      responses:
        '201':
          description: Created
          headers:
            location:
              schema:
                type: string
              description: '`/v1/subscriptions/{id}` of the new subscription.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionWithSecret'
        '400':
          description: |
            One of: `VALIDATION_FAILED` (request body fails the JSON Schema), `INVALID_URL`
            (target URL rejected by the outbound-target guard — scheme, host or private-network
            policy), `INVALID_PATTERN` (an `events` entry is not an exact type, `*`, or a
            `name.*` prefix), `INVALID_HEADER` (a custom header name/value is not allowed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                validationFailed:
                  $ref: '#/components/examples/ValidationFailedEx'
                invalidUrl:
                  $ref: '#/components/examples/InvalidUrlEx'
                invalidPattern:
                  $ref: '#/components/examples/InvalidPatternEx'
                invalidHeader:
                  $ref: '#/components/examples/InvalidHeaderEx'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/SubscriptionExists'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      operationId: webhookOut.subscriptions.list
      summary: List subscriptions
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: read
      security:
        - ApiKeyAuth: []
      parameters:
        - name: q
          in: query
          schema: { type: string, minLength: 1, maxLength: 120 }
          description: Substring match against name, description or URL.
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SubscriptionStatus' }
        - name: event
          in: query
          schema: { type: string, minLength: 1, maxLength: 120 }
          description: Only subscriptions whose `events` array contains exactly this pattern string.
        - name: limit
          in: query
          schema: { $ref: '#/components/schemas/Limit' }
        - name: cursor
          in: query
          schema: { type: string, pattern: '^[a-z0-9]+([.\-_][a-z0-9]+)*$', maxLength: 80 }
          description: Opaque keyset cursor; pass back the previous page's `nextCursor` (a subscription name).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                  nextCursor:
                    type: [string, 'null']
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/subscriptions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/SubscriptionId' }
    get:
      operationId: webhookOut.subscriptions.get
      summary: Get a subscription
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: read
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [subscription]
                properties:
                  subscription: { $ref: '#/components/schemas/Subscription' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    patch:
      operationId: webhookOut.subscriptions.patch
      summary: Update a subscription
      description: |
        Partial update; at least one field is required. `enabled: true` resumes the subscription
        (including from the automatic `disabled` state, resetting the consecutive-failure
        counter); `enabled: false` pauses it.
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSubscriptionBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [subscription]
                properties:
                  subscription: { $ref: '#/components/schemas/Subscription' }
        '400':
          description: |
            One of: `VALIDATION_FAILED`, `INVALID_URL`, `INVALID_PATTERN`, `INVALID_HEADER` — same
            meanings as on create.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                validationFailed:
                  $ref: '#/components/examples/ValidationFailedEx'
                invalidUrl:
                  $ref: '#/components/examples/InvalidUrlEx'
                invalidPattern:
                  $ref: '#/components/examples/InvalidPatternEx'
                invalidHeader:
                  $ref: '#/components/examples/InvalidHeaderEx'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '409':
          description: '`SUBSCRIPTION_EXISTS` — the patched `name` collides with another subscription.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                subscriptionExists:
                  $ref: '#/components/examples/SubscriptionExistsEx'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      operationId: webhookOut.subscriptions.delete
      summary: Delete a subscription
      description: Also deletes, by database cascade, every delivery queued or recorded for it.
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: Deleted; no content.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/subscriptions/{id}/rotate:
    post:
      operationId: webhookOut.subscriptions.rotate
      summary: Rotate a subscription's signing secret
      description: |
        Issues a brand-new signing secret immediately. **The plaintext secret is returned only in
        this response — it is never retrievable again afterward.** The previous secret keeps
        signing a second, parallel `v1` HMAC signature on outgoing deliveries until
        `previousValidUntil` (governed by `PREV_SECRET_GRACE_HOURS`; `null` when the grace period
        is configured to zero), so a receiver can roll its own verification key over without
        dropping in-flight deliveries.
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/SubscriptionId' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [subscription, secret, previousValidUntil]
                properties:
                  subscription: { $ref: '#/components/schemas/Subscription' }
                  secret:
                    $ref: '#/components/schemas/SigningSecret'
                  previousValidUntil:
                    type: [string, 'null']
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/subscriptions/{id}/test:
    post:
      operationId: webhookOut.subscriptions.test
      summary: Send a test delivery
      description: |
        Queues one synthetic `webhook.test` event for this subscription only, regardless of its
        event patterns or status (a paused subscription still gets it queued; it is sent once
        resumed). Async: `202` is a durability guarantee that the delivery is queued, not that it
        has been attempted yet.
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/SubscriptionId' }
      responses:
        '202':
          description: Accepted and queued.
          content:
            application/json:
              schema:
                type: object
                required: [event, delivery]
                properties:
                  event: { $ref: '#/components/schemas/Event' }
                  delivery: { $ref: '#/components/schemas/Delivery' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/subscriptions/{id}/replay:
    post:
      operationId: webhookOut.subscriptions.replay
      summary: Re-queue events in a time window for this subscription
      description: |
        Re-queues every already-published, non-test event in `[from, to)` that matches this
        subscription's patterns as a brand-new delivery — existing deliveries for those events are
        untouched, so the receiver will see repeated event ids and must de-duplicate itself.
        `to` defaults to now. Distinct from `test` (queues one synthetic event) and from automatic
        retry (reuses the same delivery id) — this always creates new delivery rows, one per
        matching event, and the response reports only a count, not the events or deliveries
        themselves.
      tags: [subscriptions]
      x-stack-visibility: authenticated
      x-required-role: write
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/SubscriptionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [from]
              properties:
                from:
                  $ref: '#/components/schemas/IsoDateTimeString'
                to:
                  $ref: '#/components/schemas/IsoDateTimeString'
      responses:
        '202':
          description: Accepted; every matching event has already been queued as a new delivery.
          content:
            application/json:
              schema:
                type: object
                required: [queued]
                properties:
                  queued:
                    type: integer
                    minimum: 0
                    description: Count of newly-queued deliveries, not an array of them.
              example:
                queued: 42
        '400':
          description: |
            One of: `VALIDATION_FAILED` (`from`/`to` missing or not an ISO 8601-shaped string per
            the request schema), `INVALID_RANGE` (`from`/`to` fails `Date.parse`, or `to` is not
            after `from`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                validationFailed:
                  $ref: '#/components/examples/ValidationFailedEx'
                invalidRange:
                  $ref: '#/components/examples/InvalidRangeEx'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/SubscriptionNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/events/{id}:
    get:
      operationId: webhookOut.events.get
      summary: Get an event and its deliveries
      tags: [events]
      x-stack-visibility: authenticated
      x-required-role: read
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/EventId' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [event, deliveries]
                properties:
                  event: { $ref: '#/components/schemas/Event' }
                  deliveries:
                    type: array
                    items:
                      $ref: '#/components/schemas/Delivery'
                    description: Every delivery this event ever produced, across all subscriptions.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EventNotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/stats:
    get:
      operationId: webhookOut.stats.get
      summary: Aggregate subscription/event/delivery/worker statistics
      description: '`last24h`/`recentFailures`/`avgDurationMs24h` are computed over a fixed trailing 24-hour window.'
      tags: [stats]
      x-stack-visibility: authenticated
      x-required-role: read
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <secret>`. Each configured key carries one role (`read`, `write`,
        `publish` or `readwrite`); see the grant-lattice table in the top-level description.
  schemas:
...
    SubscriptionWithSecret:
      type: object
      required: [subscription, secret]
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
        secret:
          $ref: '#/components/schemas/SigningSecret'
...
  examples:
...
    InvalidUrlEx:
      summary: INVALID_URL
      value:
        error:
          code: INVALID_URL
          message: target URL must use http or https
...